Enabled decimal input
The input type for all mass fields has been changed from `number` to `numberDecimal` to allow for decimal values.

The persistence logic in `SharedViewModel` has been updated to store and retrieve these values as strings instead of floats.

In `ScrollingFragment`, input fields are now cleared if their value is 0, improving the user experience.